fix(repo): satisfy agent registry shape checks
