+++ scripts/max_lines_allowlist.txt
@@ -28,11 +28,13 @@ apps/desktop/src/components/workspace/shell/right-panel/RightPanel.test.tsx 733 
 apps/desktop/src/lib/domain/workspaces/cloud/logical-workspaces.test.ts 662 existing desktop oversized test file
 apps/desktop/src/lib/workflows/mcp/connector-persistence.test.ts 704 existing desktop oversized test file
 server/proliferate/server/cloud/commands/domain/payload.py 530 existing cloud command payload validator with workspace materialization policies
+server/proliferate/server/cloud/agent_auth/models.py 535 existing cloud agent auth API schema module pending request/response split
+server/proliferate/server/cloud/agent_auth/runtime_keys.py 671 existing cloud agent auth runtime key service pending gateway/router split
 server/tests/integration/schema_migration_assertions.py 878 existing server oversized test helper
 server/tests/integration/test_auth_flow.py 2057 existing server oversized integration test
 server/tests/integration/test_billing_accounting.py 872 existing server oversized integration test
 server/tests/integration/test_billing_api.py 2054 existing server oversized integration test
-server/tests/integration/test_cloud_agent_auth_api.py 1664 existing server oversized integration test
+server/tests/integration/test_cloud_agent_auth_api.py 1862 existing server oversized integration test
 server/tests/integration/test_cloud_api.py 2346 existing server oversized integration test
 server/tests/integration/test_cloud_commands_api.py 4889 existing server oversized integration test
 server/tests/integration/test_sandbox_profile_foundation.py 1222 existing server oversized integration test
@@ -40,4 +42,4 @@ server/tests/integration/test_schema_migrations.py 925 existing server oversized
 server/tests/integration/test_stripe_webhooks.py 1456 existing server oversized integration test
 server/tests/unit/test_cloud_mobility_service.py 800 existing server oversized unit test
 server/tests/unit/test_cloud_runtime_ensure_running.py 872 existing server oversized unit test
-server/tests/unit/test_cloud_runtime_provision.py 1822 existing server oversized unit test
+server/tests/unit/test_cloud_runtime_provision.py 1823 existing server oversized unit test
